requirements.txt
alembic==1.7.4
autopep8==1.6.0
click==8.0.3
ecdsa==0.17.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2
zipp==3.6.0
